+++ changelog.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2021-04-30
+### Removed
+- Email admin button from admin tab
+- Email admin function from main.py
+
 ## [0.0.2] - 2021-04-28
 ### Changed
 - Changed rule_on_click to only update the list item icon instead of removing the list and regenerating.